requirements.txt
paho-mqtt
pydantic==1.10.14
pymongo==4.3.3
python-dotenv
pyspark==3.5.0